+++ api/trading_test.go
@@ -287,6 +287,9 @@ func getTestGRPCServer(
 	sqlVoteStore := sqlstore.NewVotes(&sqlStore)
 	sqlRiskFactorsStore := sqlstore.NewRiskFactors(&sqlStore)
 	sqlMarginLevelsStore := sqlstore.NewMarginLevels(&sqlStore)
+	sqlNetParamStore := sqlstore.NewNetworkParameters(&sqlStore)
+	sqlBlockStore := sqlstore.NewBlocks(&sqlStore)
+	sqlCheckpointStore := sqlstore.NewCheckpoints(&sqlStore)
 
 	g := api.NewGRPCServer(
 		logger,
@@ -335,6 +338,9 @@ func getTestGRPCServer(
 		sqlVoteStore,
 		sqlRiskFactorsStore,
 		sqlMarginLevelsStore,
+		sqlNetParamStore,
+		sqlBlockStore,
+		sqlCheckpointStore,
 	)
 	if g == nil {
 		err = fmt.Errorf("failed to create gRPC server")