+++ cmd/data-node/node/node_pre.go
@@ -140,6 +140,8 @@ func (l *NodeCommand) setupSQLSubscribers() {
 	l.votesSubSQL = sqlsubscribers.NewVote(l.voteStoreSQL, l.Log)
 	l.marginLevelsSubSQL = sqlsubscribers.NewMarginLevels(l.marginLevelsStoreSQL, l.Log)
 	l.riskFactorSubSQL = sqlsubscribers.NewRiskFactor(l.riskFactorStoreSQL, l.Log)
+	l.netParamSubSQL = sqlsubscribers.NewNetworkParameter(l.netParamStoreSQL, l.Log)
+	l.checkpointSubSQL = sqlsubscribers.NewCheckpoint(l.checkpointStoreSQL, l.Log)
 }
 
 func (l *NodeCommand) setupStorages() error {
@@ -184,6 +186,8 @@ func (l *NodeCommand) setupStorages() error {
 		l.voteStoreSQL = sqlstore.NewVotes(sqlStore)
 		l.marginLevelsStoreSQL = sqlstore.NewMarginLevels(sqlStore)
 		l.riskFactorStoreSQL = sqlstore.NewRiskFactors(sqlStore)
+		l.netParamStoreSQL = sqlstore.NewNetworkParameters(sqlStore)
+		l.checkpointStoreSQL = sqlstore.NewCheckpoints(sqlStore)
 
 		l.sqlStore = sqlStore
 	}
@@ -272,7 +276,10 @@ func (l *NodeCommand) preRun(_ []string) (err error) {
 			l.votesSubSQL,
 			l.depositSubSQL,
 			l.marginLevelsSubSQL,
-			l.riskFactorSubSQL)
+			l.riskFactorSubSQL,
+			l.netParamSubSQL,
+			l.checkpointSubSQL,
+		)
 	}
 
 	l.broker, err = broker.New(l.ctx, l.Log, l.conf.Broker, l.chainInfoStore, eventSource)