+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - [435](https://github.com/vegaprotocol/data-node/pull/435) - Migrate governance API to retrieve data from `Postgres`
 - [442](https://github.com/vegaprotocol/data-node/pull/442) - Migrate estimator API to retrieve data from `Postgres`
 - [449](https://github.com/vegaprotocol/data-node/issue/449) - Refactor identifiers to use `ID` types instead of `[]byte`
+- [373](https://github.com/vegaprotocol/data-node/issue/373) - Migrate `GetVegaTime`, `Checkpoints` and `NetworkParameters` to `Postgres`
 
 ### 🐛 Fixes
 - [256](https://github.com/vegaprotocol/data-node/pull/256) - Market Risk Factors missing from Market `GraphQL` API